fix issue Incorrect decoding of FreshnessPeriod #33

+++ encode/metainfo.h
@@ -26,7 +26,7 @@ typedef struct ndn_metainfo {
   /**
    * The freshness period of the Data packet.
    */
-  uint32_t freshness_period;
+  uint64_t freshness_period;
   /**
    * The last name component in Name.
    */
@@ -97,7 +97,7 @@ ndn_metainfo_set_content_type(ndn_metainfo_t* meta, uint8_t content_type)
  * @param freshness_period. Input. FreshnessPeriod value.
  */
 static inline void
-ndn_metainfo_set_freshness_period(ndn_metainfo_t* meta, uint32_t freshness_period)
+ndn_metainfo_set_freshness_period(ndn_metainfo_t* meta, uint64_t freshness_period)
 {
   meta->enable_FreshnessPeriod = 1;
   meta->freshness_period = freshness_period;
